moved StringUtil into StrUtil to avoid conflicts with another file with lower case name in gui/util
