+++ retroshare-gui/src/TorControl/StrUtil.h
